+++ tests/base/model_utilities.py
@@ -26,25 +26,3 @@ class ModelTemplateUtils:
         else:  # if it is 2level deep -> per dataloader and per batch
             val = sum(out[name] for out in output) / len(output)
         return val
-
-
-class CustomInfDataloader:
-
-    def __init__(self, dataloader):
-        self.dataloader = dataloader
-        self.iter = iter(dataloader)
-        self.count = 0
-
-    def __iter__(self):
-        self.count = 0
-        return self
-
-    def __next__(self):
-        if self.count >= 50:
-            raise StopIteration
-        self.count = self.count + 1
-        try:
-            return next(self.iter)
-        except StopIteration:
-            self.iter = iter(self.dataloader)
-            return next(self.iter)
