+++ tests/test_deprecated.py
@@ -6,7 +6,7 @@ import pytest
 from pytorch_lightning import Trainer
 
 import tests.base.utils as tutils
-from tests.base import TestModelBase, LightTrainDataloader, LightEmptyTestStep
+from tests.base import EvalModelTemplate
 
 
 def _soft_unimport_module(str_module):
@@ -120,11 +120,11 @@ def test_tbd_remove_in_v0_9_0_module_imports():
         from pytorch_lightning.logging.wandb import WandbLogger  # noqa: F402
 
 
-class ModelVer0_6(LightTrainDataloader, LightEmptyTestStep, TestModelBase):
+class ModelVer0_6(EvalModelTemplate):
 
     # todo: this shall not be needed while evaluate asks for dataloader explicitly
     def val_dataloader(self):
-        return self._dataloader(train=False)
+        return self.dataloader(train=False)
 
     def validation_step(self, batch, batch_idx, *args, **kwargs):
         return {'val_loss': 0.6}
@@ -133,17 +133,17 @@ class ModelVer0_6(LightTrainDataloader, LightEmptyTestStep, TestModelBase):
         return {'val_loss': 0.6}
 
     def test_dataloader(self):
-        return self._dataloader(train=False)
+        return self.dataloader(train=False)
 
     def test_end(self, outputs):
         return {'test_loss': 0.6}
 
 
-class ModelVer0_7(LightTrainDataloader, LightEmptyTestStep, TestModelBase):
+class ModelVer0_7(EvalModelTemplate):
 
     # todo: this shall not be needed while evaluate asks for dataloader explicitly
     def val_dataloader(self):
-        return self._dataloader(train=False)
+        return self.dataloader(train=False)
 
     def validation_step(self, batch, batch_idx, *args, **kwargs):
         return {'val_loss': 0.7}
@@ -152,7 +152,7 @@ class ModelVer0_7(LightTrainDataloader, LightEmptyTestStep, TestModelBase):
         return {'val_loss': 0.7}
 
     def test_dataloader(self):
-        return self._dataloader(train=False)
+        return self.dataloader(train=False)
 
     def test_end(self, outputs):
         return {'test_loss': 0.7}
